Add a prctl to disable path traversal
https://infosec.exchange/@joshbressers/111840317513981607 makes sense.
Let's add an opt-in prctl, and then systemd or whatever can be used to
set this. There's a somewhat hacky workaround to permit resolution of
symlinks that point to parent directories (so, eg, the standard pattern
of ln -s ../sites-available/foo sites-enabled/foo works). I'm sure it
breaks in other exciting ways, though.

Signed-off-by: Matthew Garrett <mgarrett@aurora.tech>